+++ frontend/src/App.tsx
@@ -5,6 +5,8 @@ import RevisionPlan from './pages/RevisionPlan'
 import QCM from './pages/QCM'
 import QCMResult from './pages/QCMResult'
 import Dashboard from './pages/Dashboard'
+import Simulation from './pages/Simulation'
+import QCMGenere from './pages/QCMGenere'
 
 const ExerciceRedige = lazy(() => import('./pages/ExerciceRedige'))
 
@@ -19,6 +21,8 @@ function App() {
           <Route path="/qcm/:exerciseId/result" element={<QCMResult />} />
           <Route path="/redige/:exerciseId" element={<ExerciceRedige />} />
           <Route path="/dashboard" element={<Dashboard />} />
+          <Route path="/simulation" element={<Simulation />} />
+          <Route path="/qcm-genere" element={<QCMGenere />} />
         </Routes>
       </Suspense>
     </BrowserRouter>